+++ services/apps/cache_worker/src/workflows.ts
@@ -1,11 +1,11 @@
-import { computeOrgAggsAndUpdate } from './workflows/compute-orgs-agg/computeOrgAggsAndUpdate'
-import { dailyGetAndComputeOrgAggs } from './workflows/compute-orgs-agg/getAndComputeOrgAggs'
+import { processOrganizationAggregates } from './workflows/processOrganizationAggregates'
 import { refreshDashboardCache } from './workflows/refreshDashboardCache'
 import { spawnDashboardCacheRefresh } from './workflows/spawnDashboardCacheRefresh'
+import { spawnOrganizationAggregatesComputation } from './workflows/spawnOrganizationAggregatesComputation'
 
 export {
   spawnDashboardCacheRefresh,
   refreshDashboardCache,
-  dailyGetAndComputeOrgAggs,
-  computeOrgAggsAndUpdate,
+  spawnOrganizationAggregatesComputation,
+  processOrganizationAggregates,
 }
